+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+
+- Arguments:
+  - '--confirma-category'
+
 ## [0.8.0-beta 2024-08-18]
 
 ### Added